handle redlinks
this is using an adapted version of format.js, again.

we use ajax HEAD requests to the /raw route (as non-existant links will 302 to /edit on /wiki)

a cache is used, powered by a key-value store of an object to avoid unneccesary requests to the server. i have a slight concern it might leak if you dawdle too much typing in a CamelCase link.