aggresive refactor, part 2
just did a res\.sendStatus\(\d{3}\)\.end() regex and added a return to each, even if not explicitly needed. better safe than sorry.

i removed the synchrousity of the AJAX call on editor.js due to deadlocking if the server crashed. this unfortunately causes a race condition where the link is made before the return of the page state, but it will update fine next character.